zephyr: Fix ping fail issue after 11r roaming finishes
Ping fail issue was found after 11r roaming finishes. This issue is
caused by incorrect key. When doing 11r roaming, wpa_supplicant
calculated new PTK of AP2 and send it to driver. In this case, STA is
still associated to AP1 and the key installation might fail. To fix
this issue, send new PTK and GTK after associating with AP2.

Signed-off-by: Hui Bai <hui.bai@nxp.com>