zephyr: udp throughput fail when legacy roaming
zperf udp tx throughput uses fixed ip address, if run legacy roaming at
this time, roaming will restart dhcp and ip address is changed.
so zperf can't find ip address when zperf send udp packet, and zperf udp
tx fail.
we think that legacy roaming shouldn't restart dhcp, so we need to
determine whether current connection is a normal connection or a
roaming, and determine whether we should restart dhcp.
and we use assoc_freq parameter in wps_supplicant to get current connect
state.
if assoc_freq is not 0, it indicates that the sta already connects, and
current connection is roaming, so we doesn't start dhcp.
if assoc_freq is 0, it indicates that current connection is a normal
connection rather than roaming, so we want to start dhcp.

Signed-off-by: Gaofeng Zhang <gaofeng.zhang@nxp.com>