Fix multiple size_t print format specifiers
The format specifier for size_t is zu.
Using d only works when int and size_t are the same
underlying type which is not the case for 64bit systems,
which leads to a build warning in this case.

In 2 places, instead of changing the printf specifier
the type of the variable was changed to int, as that
was what the code really exepected.
And while at it, two internal function have been made static.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>