[fromtree] wlantest: Use AP's RSNXOE for capabilities when RSNO is used
If an association uses RSN overriding and the AP advertises an RSNXOE,
use the RSNXOE instead of the RSNXE when determining AP's RSN
capabilities. In particular, this is needed to determine the correct KDK
length for PTK derivation in a case where the RSNXOE advertises support
for secure ranging while the RSNXE does not.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
(cherry picked from commit 59c608bd8d55089ef8d4aa2f2b7dd46ddd2e4f31)